refactor: make Task and Concurrent mixins broker-agnostic
