+++ faststream/nats/subscriber/usecase.py
@@ -1,4 +1,3 @@
-import asyncio
 from abc import abstractmethod
 from contextlib import suppress
 from typing import (
@@ -6,7 +5,6 @@ from typing import (
     Any,
     Awaitable,
     Callable,
-    Coroutine,
     Dict,
     Generic,
     Iterable,
@@ -25,6 +23,7 @@ from nats.js.api import ConsumerConfig, ObjectInfo
 from typing_extensions import Annotated, Doc, override
 
 from faststream.broker.publisher.fake import FakePublisher
+from faststream.broker.subscriber.mixins import ConcurrentMixin, TasksMixin
 from faststream.broker.subscriber.usecase import SubscriberUsecase
 from faststream.broker.types import MsgType
 from faststream.broker.utils import process_msg
@@ -44,7 +43,6 @@ from faststream.nats.subscriber.subscription import (
 from faststream.utils.context.repository import context
 
 if TYPE_CHECKING:
-    from anyio.streams.memory import MemoryObjectReceiveStream, MemoryObjectSendStream
     from nats.aio.client import Client
     from nats.aio.msg import Msg
     from nats.aio.subscription import Subscription
@@ -320,73 +318,6 @@ class _DefaultSubscriber(LogicSubscriber[ConnectionType, MsgType]):
         )
 
 
-class _TasksMixin(LogicSubscriber[Any, Any]):
-    def __init__(self, **kwargs: Any) -> None:
-        self.tasks: List[asyncio.Task[Any]] = []
-
-        super().__init__(**kwargs)
-
-    def add_task(self, coro: Coroutine[Any, Any, Any]) -> None:
-        self.tasks.append(asyncio.create_task(coro))
-
-    async def close(self) -> None:
-        """Clean up handler subscription, cancel consume task in graceful mode."""
-        await super().close()
-
-        for task in self.tasks:
-            if not task.done():
-                task.cancel()
-
-        self.tasks = []
-
-
-class _ConcurrentMixin(_TasksMixin):
-    send_stream: "MemoryObjectSendStream[Msg]"
-    receive_stream: "MemoryObjectReceiveStream[Msg]"
-
-    def __init__(
-        self,
-        *,
-        max_workers: int,
-        **kwargs: Any,
-    ) -> None:
-        self.max_workers = max_workers
-
-        self.send_stream, self.receive_stream = anyio.create_memory_object_stream(
-            max_buffer_size=max_workers
-        )
-        self.limiter = anyio.Semaphore(max_workers)
-
-        super().__init__(**kwargs)
-
-    def start_consume_task(self) -> None:
-        self.add_task(self._serve_consume_queue())
-
-    async def _serve_consume_queue(
-        self,
-    ) -> None:
-        """Endless task consuming messages from in-memory queue.
-
-        Suitable to batch messages by amount, timestamps, etc and call `consume` for this batches.
-        """
-        async with anyio.create_task_group() as tg:
-            async for msg in self.receive_stream:
-                tg.start_soon(self._consume_msg, msg)
-
-    async def _consume_msg(
-        self,
-        msg: "Msg",
-    ) -> None:
-        """Proxy method to call `self.consume` with semaphore block."""
-        async with self.limiter:
-            await self.consume(msg)
-
-    async def _put_msg(self, msg: "Msg") -> None:
-        """Proxy method to put msg into in-memory queue with semaphore block."""
-        async with self.limiter:
-            await self.send_stream.send(msg)
-
-
 class CoreSubscriber(_DefaultSubscriber["Client", "Msg"]):
     subscription: Optional["Subscription"]
     _fetch_sub: Optional["Subscription"]
@@ -499,7 +430,7 @@ class CoreSubscriber(_DefaultSubscriber["Client", "Msg"]):
 
 
 class ConcurrentCoreSubscriber(
-    _ConcurrentMixin,
+    ConcurrentMixin,
     CoreSubscriber,
 ):
     def __init__(
@@ -692,7 +623,7 @@ class PushStreamSubscription(_StreamSubscriber):
 
 
 class ConcurrentPushStreamSubscriber(
-    _ConcurrentMixin,
+    ConcurrentMixin,
     _StreamSubscriber,
 ):
     subscription: Optional["JetStreamContext.PushSubscription"]
@@ -760,7 +691,7 @@ class ConcurrentPushStreamSubscriber(
 
 
 class PullStreamSubscriber(
-    _TasksMixin,
+    TasksMixin,
     _StreamSubscriber,
 ):
     subscription: Optional["JetStreamContext.PullSubscription"]
@@ -845,7 +776,7 @@ class PullStreamSubscriber(
 
 
 class ConcurrentPullStreamSubscriber(
-    _ConcurrentMixin,
+    ConcurrentMixin,
     PullStreamSubscriber,
 ):
     def __init__(
@@ -910,7 +841,7 @@ class ConcurrentPullStreamSubscriber(
 
 
 class BatchPullStreamSubscriber(
-    _TasksMixin,
+    TasksMixin,
     _DefaultSubscriber["JetStreamContext", List["Msg"]],
 ):
     """Batch-message consumer class."""
@@ -990,12 +921,12 @@ class BatchPullStreamSubscriber(
         except TimeoutError:
             return None
 
-        msg: NatsMessage = await process_msg(
+        msg = cast(NatsMessage, await process_msg(
             msg=raw_message,
             middlewares=self._broker_middlewares,
             parser=self._parser,
             decoder=self._decoder,
-        )
+        ))
         return msg
 
     @override
@@ -1031,7 +962,7 @@ class BatchPullStreamSubscriber(
 
 
 class KeyValueWatchSubscriber(
-    _TasksMixin,
+    TasksMixin,
     LogicSubscriber["KVBucketDeclarer", "KeyValue.Entry"],
 ):
     subscription: Optional["UnsubscribeAdapter[KeyValue.KeyWatcher]"]
@@ -1188,7 +1119,7 @@ OBJECT_STORAGE_CONTEXT_KEY = "__object_storage"
 
 
 class ObjStoreWatchSubscriber(
-    _TasksMixin,
+    TasksMixin,
     LogicSubscriber["OSBucketDeclarer", ObjectInfo],
 ):
     subscription: Optional["UnsubscribeAdapter[ObjectStore.ObjectWatcher]"]
